stable entry using cf assets
