+++ eslint.config.mjs
@@ -37,6 +37,7 @@ const opts = tseslint.config(
       "vibesbox/**",
       "vibes.diy/tempo/**",
       "vibes.diy/failback-homepage/**",
+      "vibes.diy/stable-entry/public/**",
       "hosting/**",
       "**/coverage/**",
       "**/.cache/**",
