+++ pnpm-workspace.yaml
@@ -18,6 +18,7 @@ packages:
   - "vibes.diy/tests/app"
   - "vibes.diy/tests/simple-chat"
   - "vibes.diy/failback-homepage"
+  - "vibes.diy/stable-entry"
   - "prompts/pkg"
   - "prompts/tests"
   - "hosting/base"
