+++ vibes.diy/stable-entry/package.json
@@ -6,7 +6,11 @@
     "deploy": "wrangler deploy",
     "dev": "wrangler dev"
   },
+  "dependencies": {
+    "arktype": "^2.2.0"
+  },
   "devDependencies": {
+    "@cloudflare/workers-types": "^4.20260317.1",
     "wrangler": "^4.74.0"
   }
 }
