+++ vibes.diy/stable-entry/public/.stable-entry/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>stable-entry</title>
+  <script type="importmap">
+  {
+    "imports": {
+      "react": "https://esm.sh/react@18?bundle",
+      "react-dom/client": "https://esm.sh/react-dom@18/client?external=react&bundle"
+    }
+  }
+  </script>
+</head>
+<body>
+  <div id="root"></div>
+  <script type="module" src="/.stable-entry/app.js"></script>
+</body>
+</html>
