+++ vibes.diy/stable-entry/src/index.ts
@@ -1,11 +1,12 @@
+import { type } from "arktype";
+
 interface Env {
   BACKEND: string;
   BACKENDS?: string;
 }
 
-interface BackendsMap {
-  [key: string]: string;
-}
+const backendsType = type("Record<string, string>");
+type BackendsMap = typeof backendsType.infer;
 
 let cachedBackends: BackendsMap | null = null;
 let cachedBackendsRaw: string | undefined;
@@ -14,20 +15,14 @@ function parseBackends(raw: string | undefined): BackendsMap | null {
   if (!raw) return null;
   if (raw === cachedBackendsRaw && cachedBackends !== null) return cachedBackends;
   try {
-    const parsed = JSON.parse(raw);
-    if (typeof parsed !== "object" || parsed === null || Array.isArray(parsed)) {
-      console.error("BACKENDS must be a JSON object, got:", typeof parsed);
+    const result = backendsType(JSON.parse(raw));
+    if (result instanceof type.errors) {
+      console.error("BACKENDS validation failed:", result.summary);
       return null;
     }
-    for (const [k, v] of Object.entries(parsed)) {
-      if (typeof v !== "string") {
-        console.error(`BACKENDS["${k}"] must be a string URL, got:`, typeof v);
-        return null;
-      }
-    }
     cachedBackendsRaw = raw;
-    cachedBackends = parsed as BackendsMap;
-    return cachedBackends;
+    cachedBackends = result;
+    return result;
   } catch (e) {
     console.error("Failed to parse BACKENDS JSON:", e);
     return null;
@@ -98,111 +93,30 @@ function sanitizeResponseHeaders(response: Response, requestUrl: URL, backendOri
   });
 }
 
-function serveUI(backends: BackendsMap | null): Response {
-  const keys = backends ? Object.keys(backends) : [];
-  const html = `<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>stable-entry</title>
-  <script type="importmap">
-  {
-    "imports": {
-      "react": "https://esm.sh/react@18?bundle",
-      "react-dom/client": "https://esm.sh/react-dom@18/client?external=react&bundle"
-    }
-  }
-  </script>
-</head>
-<body>
-  <div id="root"></div>
-  <script type="module">
-    import React from "react";
-    import { createRoot } from "react-dom/client";
-
-    const h = React.createElement;
-    const KEYS = ${JSON.stringify(keys)};
-
-    function getCookie(name) {
-      for (const part of document.cookie.split(";")) {
-        const [k, ...rest] = part.trim().split("=");
-        if (k === name) return rest.join("=");
-      }
-      return "";
-    }
-
-    function App() {
-      const current = getCookie("Vibes-Backend") || "(default)";
-
-      function select(key) {
-        if (key) {
-          document.cookie = "Vibes-Backend=" + key + "; Path=/; SameSite=Lax; Max-Age=86400";
-        } else {
-          document.cookie = "Vibes-Backend=; Path=/; SameSite=Lax; Max-Age=0";
-        }
-        window.location.href = "/";
-      }
-
-      return h("div", { style: { fontFamily: "system-ui", maxWidth: "400px", margin: "80px auto", padding: "20px" } },
-        h("h2", null, "stable-entry"),
-        h("p", null, "Current backend: ", h("strong", null, current)),
-        h("div", null,
-              KEYS.length > 0
-                ? h("div", null,
-                    h("p", null, "Select a backend:"),
-                    h("div", { style: { display: "flex", flexDirection: "column", gap: "8px" } },
-                      KEYS.map(function(key) {
-                        return h("button", {
-                          key: key,
-                          onClick: function() { select(key); },
-                          style: {
-                            padding: "8px 16px",
-                            cursor: "pointer",
-                            fontWeight: current === key ? "bold" : "normal",
-                            border: current === key ? "2px solid #333" : "1px solid #ccc",
-                            borderRadius: "4px",
-                            background: current === key ? "#f0f0f0" : "#fff"
-                          }
-                        }, key);
-                      })
-                    )
-                  )
-                : h("p", { style: { color: "#888" } }, "No alternate backends configured."),
-              current !== "(default)"
-                ? h("button", {
-                    onClick: function() { select(""); },
-                    style: {
-                      padding: "8px 16px",
-                      cursor: "pointer",
-                      border: "1px solid #ccc",
-                      borderRadius: "4px",
-                      background: "#fff",
-                      marginTop: "8px"
-                    }
-                  }, "Reset to default")
-                : null
-            )
-      );
-    }
-
-    createRoot(document.getElementById("root")).render(h(App));
-  </script>
-</body>
-</html>`;
-
-  return new Response(html, {
-    headers: { "Content-Type": "text/html; charset=utf-8" },
-  });
-}
-
 export default {
   async fetch(request: Request, env: Env): Promise<Response> {
     const url = new URL(request.url);
 
-    if (url.pathname.startsWith("/.stable-entry")) {
+    if (url.pathname === "/.stable-entry/config.json") {
       const backends = parseBackends(env.BACKENDS);
-      return serveUI(backends);
+      const keys = backends ? Object.keys(backends) : [];
+      return new Response(JSON.stringify({ keys }), {
+        headers: { "Content-Type": "application/json" },
+      });
+    }
+
+    const backendParam = url.searchParams.get("backend");
+    if (backendParam !== null) {
+      const backends = parseBackends(env.BACKENDS);
+      url.searchParams.delete("backend");
+      const headers = new Headers();
+      headers.set("Location", url.pathname + url.search);
+      if (backendParam && backends && Object.hasOwn(backends, backendParam)) {
+        headers.set("Set-Cookie", `Vibes-Backend=${backendParam}; Path=/; SameSite=Lax; Max-Age=86400`);
+      } else {
+        headers.set("Set-Cookie", "Vibes-Backend=; Path=/; SameSite=Lax; Max-Age=0");
+      }
+      return new Response(null, { status: 302, headers });
     }
 
     const backendUrl = resolveBackendUrl(request, env);
